Added the ability to define full qualified classnames in build properties.
Propel is using dot-path notation to load a class defined in build
properties. E.g. for behaviors or builders or whatever you want.

This is great but it's better to allow full qualified classnames, to
check if it exists and to return it without any call to
`Phing::import()`.

Why? Because frameworks like Symfony2 have their own autoloader and it
could be nice to use it directly. Now you can define a class instead of
a dot-path notation and if this class exists, it will use it !

Example:

    propel.behavior.formidable.class = \My\Namespace\FormidableBehavior