Introducing the `ProfilerConnectionWrapper` class.
Move all the profiling logic out of the `ConnectionWrapper` class, so
that users who do not need profiling do not suffer from many tests in
the connection class. The logic is now better separated, the profiler is
a (reusable) service.

Note that the old way to configure profiling (through runtime-conf.xml)
no longer works. A new way to configure it from a file must still be
introduced.