Store Date and Time values as DateTime-instances in generated classes
This commit will modify the generated models so that date and time values are internally stored as DateTime instances.

Included in the commit is tests for the behavior.

This commit also takes into account the comparison of default values etc.

This commit also adds tests and handling for hydration of 0000-00-00 and 0000-00-00 00:00:00 values that are supported on MySQL.