[monolog] Made per-connection logger possible again.
As planned, this implies passing the connection name to the connection
via the ConnectionManager. It changes two interfaces, but it's not that
big a deal.
The modification also allows us to get rid of the Configuration object
in the ConnectionWrapper, so the next step is to remove this object from
the Runtime code altogether (its features are now handled mostly by the
ServiceContainer).