Fixed the Pdo MysqlAdapter dns preparement.
The MysqlAdapter while in the test and prepareParam method
did expect the charset setting to be an array and have an
additional key 'value' in which the actual character set to
be stored.

In the actual runtime-conf file, there exists no such level
of nesting, and nor does it make sens for it to be.