Fix parsing of charset by PDO at runtime
Since the removal of Phing, the runtime configuration settings are not passed the same way as before. The generated PHP configuration shows something like:

    $conf = array(
		  ...
			'settings' => array('charset' => 'utf8')
		);

while it used to be:

    $conf = array(
		  ...
		  'settings' => array('charset' => array('value' => 'utf8'))
		);

Dunno exactly when that changed, but updating to master broke the tests, and the only way to fix them is the attached changes.