Add runtime configuration converters.
This prepares the refactoring of the runtime configuration cache, and
allows the transformation of a configuration array into a php
configuration (using the ServiceContainer) to occur at buildtime rather
than runtime.

Not to mention that it clarifies the runtime configuration settings, and
encourages the direct use of PHP configuration rather than XML.

The commit also fixes a bad parsing of the <slaves> element. The
documentation explains that <slave> should be under <datasource>, while
the configuration transformation done in Propel::setConfiguration()
expected the <slave> element to be under <connection>.

Refs 181.