ModelCriteria implements IteratorAggregate
Now the query can be iterated on, and this executes a find() in the background:

```
$books = BookQuery::create()
  ->filterByTitle('War%'); // see Ma, no find()!
foreach ($books as $book) {
  echo $book->getTitle();
}
```

In addition to saving one line of code, this delays the actual query execution to the moment
it's iterated on. Using MVC frameworks, this typically happens in the view rather than in
the controller, and therefore it can be seen as a way to hit the database more lazily.