Update the contribution guideline
Since the new repository's design, the process to create a pull request
has been a little bit change ; update the documentation to match the
actual one.

Also remove the contribution part in the README file and move it to the
contributing file. The README was up to date with the new testing
process.

Also update the part about fixing checkstyle. To keep clean logs and
make git blame easier, we should only fix files which have been changed.