Fix referenceOnly bug
The propel XML schema definition allows to reference external models,
which then are included the generation of the php source code. The
resulting classes can then be used in the application.

The 'referenceOnly' parameter indicates if the model should be included
in sql schema generation or not.

The current implementation does not check the value of this parameter,
it checks only if it is set. Because of a default value, the current
check evaluates always to true. This causes the sql generation to be
skipped.