Fix the name of `filterByXxx` methods and phpName of properties
Before this commit, if a property name contained an underscore,
the name of the method wasn't converted in camel case, but only
uppercased. E.g. `left_value` generated a `filterByLeft_value`
method. Now, the previous property name generates correctly the
`filterByLeftValue` method.
According to previous Propel versions, the phpName of a property is
the upper camel case of the name. E.g. the phpName of `left_value`
property is `LeftValue`.