Naming strategy
Introduce a more strict and coherent naming strategy. Propel object
model classes now have only two properties describing the name of
the object: `name` and `sqlName`.
`name` property is the camelCase object name, to be used in PHP code.
`sqlName` property is the real database object name, to be used in
sql statements.