Support XML Inclusions in configuration, adds #1321
Implemented via the existing functionality in PHPs' DOMDocument object.

Additionally libxml_disable_entity_loader(true) is not necessary any
longer (it prevented opening XML files from disc earlier) since Oct 2012
with the release of libxml2 version 2.9.0.

The old code looked influenced by the (earlier to that date) hotfix in
the Symfony project however the XXE attack Symfony was prone to, did not
apply to the XmlToArrayConverter (no validate on parse).

Refs:

- #1321

- https://www.w3.org/TR/xinclude/

- https://mail.gnome.org/archives/xml/2012-October/msg00045.html