Add `withXyzQuery` function to aid static type checking
The `withXyzQuery` function neatly wraps the calls to `useQuery` and `endUse`
in a single function call, taking a callable working on the correct
subquery type.

Projects using propel2 can use it like this:
```php
$bookQuery
    ->withAuthorQuery(
        fn (AuthorQuery $q) => $q->filterByFirstName('Lev')
    )
    ->filterByTitle('Anna Karenina');
```

without type checkers like phpstan and psalm getting confused about the
actual return type of `endUse()`.

An extra `@phpstan-param` phpdoc line is provided for static type
checkers to check whether the argument to the callable and the return
type actually match.